requirements.txt
flask
werkzeug
openai-whisper
gunicorn
numpy
